kryten-2.1.7  release.
Introduced the --monitor, -m option that allows the configuration "file" to be
specified as as single command line argument to kryten.

This is mainly to support the command line configuration specification, multiple
channel monitoring specifications can be specified on one line, separated by a
semi-colon (;) symbol.

Added a builit in quit command that terminates kryten. This takes an option integer
parameter that can be used as the exit code.

A channel monitoring specification value can now be qualified with ">=", ">", "<",
"<=" and "/=". The previously implied "=" qualifier may now also be explicit.

Exit codes from sig int and sig term are now the same as if these signals were
not caught.